:sparkles: [BE_DH] - event list get 구현

+++ backend/src/main/java/com/ssafy/pit/repository/ClassRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.pit.entity.Classes;
 
 @Repository
-public interface ClassRepository extends JpaRepository<Classes, Long> {
+public interface ClassRepository extends JpaRepository<Classes, Integer> {
 //	User findUserByUserEmail(String userEmail);
 //	User findUserByUserNickname(String userNickname);
 //	List<Classes> getAllClasses();
